requirements.txt
Django==1.6.2
Pillow==2.4.0
PyPDF2==1.20
argparse==1.2.1
distribute==0.6.24
django-easy-pdf==0.1.0
html5lib==0.999
pisa==3.0.33
pyPdf==1.13
reportlab==2.7
six==1.6.1
wsgiref==0.1.2
xhtml2pdf==0.0.5
